move arg parsing into separate file
