+++ lib/index.js
@@ -1,6 +1,7 @@
 var request = require('request');
 var _ = require('lodash');
 var Q = require('q');
+var argParse = require('argparse');
 
 var DEFAULT_LIMIT = 500;
 var debug = _.identity;
@@ -10,7 +11,7 @@ var getOpts = function (q, options) {
     url: options.baseUrl + '/api/items/search.json',
     qs: {
       q: q,
-      limit: options.limit || DEFAULT_LIMIT
+      limit: options.LIMIT || DEFAULT_LIMIT
     },
     auth: {
       username: options.user,
@@ -28,16 +29,10 @@ var search = function (opts) {
     debug("First response: ", (new Date()) - start, 'ms');
     var total = JSON.parse(body).total_count;
     debug('total items: ', total);
-    var strs = [
-      [null, body]
-    ];
+    var strs = [];
     var pages = Math.ceil(total / opts.qs.limit);
-    
+
     _.times(pages, function (pageNumber) {
-      if (pageNumber === 0) {
-        // We've already processed page 1.
-        return;
-      }
       opts.qs.offset = pageNumber * opts.qs.limit;
       strs.push(Q.nfcall(request.get, opts));
     });
@@ -61,53 +56,7 @@ var search = function (opts) {
 
 /* istanbul ignore if  */
 if (require.main === module) {
-
-  var yargs = require('yargs');
-  var args = yargs
-    .demand(['user', 'key'])
-    .describe('user', 'Email address you log in with')
-    .describe('key', 'API key (it\'s on your profile page)')
-    .describe('debug', 'Enable debug mode for extra reporting')
-    .describe('query', 'Search query.')
-    .default('query', 'assigned_to:me')
-    .describe('base_url', 'Root URL to hit for sprintly endpoint')
-    .default('base_url', 'https://sprint.ly')
-    .describe('help', 'Show this help')
-    .describe('limit', 'API limit to use')
-    .default('limit', DEFAULT_LIMIT)
-    .describe('nojson', "Don't print json. Used in combination with --debug")
-    .default('nojson', false)
-    .argv;
-
-  var removeTrailingSlash = /(.*)\/$/g;
-  var baseUrl = removeTrailingSlash.exec(args.base_url);
-  if (baseUrl) {
-    baseUrl = baseUrl[1];
-  } else {
-    baseUrl = args.base_url;
-  }
-  
-  if (args.help) {
-    console.log(yargs.help());
-    return;
-  }
-
-  if (args.debug) {
-    debug = console.log;
-  }
-
-  search(getOpts(args.query, {
-    user: args.user,
-    pass: args.key,
-    baseUrl: baseUrl,
-    limit: args.limit
-  })).then(function (json) {
-    if (!args.nojson) {
-      console.log(JSON.stringify(json));
-    }
-  }, function (err) {
-    console.log('error: ', err.response.body);
-  });
+  argParse();
 }
 
 module.exports.search = search;
